+++ src/builtin/module_builtin_array.cpp
@@ -93,12 +93,13 @@ namespace das {
     }
 
     void builtin_array_erase_range_i64 ( Array & pArray, int64_t index, int64_t count, int stride, Context * context, LineInfoArg * at ) {
-        if ( index < 0 || count < 0 || uint64_t(index + count) > pArray.size ) {
+        // Compute end as uint64 sum AFTER non-negativity check to avoid signed overflow UB on index+count.
+        if ( index < 0 || count < 0 || uint64_t(index) + uint64_t(count) > pArray.size ) {
             context->throw_error_at(at, "erasing array range is invalid: index=%lld count=%lld size=%llu", (long long)index, (long long)count, (unsigned long long)pArray.size);
             return;
         }
-        memmove ( pArray.data+index*stride, pArray.data+(index+count)*stride, size_t(pArray.size-uint64_t(index)-uint64_t(count))*size_t(stride) );
-        array_resize(*context, pArray, pArray.size-count, stride, false, at);
+        memmove ( pArray.data+uint64_t(index)*stride, pArray.data+(uint64_t(index)+uint64_t(count))*stride, size_t(pArray.size-uint64_t(index)-uint64_t(count))*size_t(stride) );
+        array_resize(*context, pArray, pArray.size-uint64_t(count), stride, false, at);
     }
 
     void builtin_array_clear ( Array & pArray, Context * context, LineInfoArg * at ) {
